+++ src/app/store-stats/AccountCurrencies.m.scss
@@ -7,7 +7,3 @@
   overflow: hidden;
   text-overflow: ellipsis;
 }
-
-.faded {
-  opacity: 0.7;
-}
