+++ src/app/store-stats/AccountCurrencies.m.scss.d.ts
@@ -1,7 +1,6 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'faded': string;
   'icon': string;
   'text': string;
 }
